# A more low level implementation of vectorize in numba