… pytensor/link/numba/dispatch/helpers.py
… pytensor/link/numba/linker.py
… tests/link/numba/test_basic.py
… tests/link/numba/test_elemwise.py
… tests/link/numba/test_extra_ops.py
… tests/link/numba/test_scalar.py