As discussed in https://github.com/pymc-devs/pytensor/discussions/70, this rewrites the numba implementation using llvm intrinsics.
This way we can use broadcasting information, supply llvm with required aliasing information for vectoriziation, support multiple outputs of elemwise and prepare for ElemwiseSum-like optimizations. (the llvm code supports that already).

Locally, there are still some test failures due to execution in python mode that I will have to work out.

This can lead to sizable performance improvements (with [svml](https://numba.readthedocs.io/en/stable/user/performance-tips.html#intel-svml) installed):

```python
import pytensor
import pytensor.tensor as pt
import numpy as np
import numba
import pytensor.link.numba.dispatch

import llvmlite.binding as llvm
#llvm.set_option('', '--debug-only=loop-vectorize,iv-descriptors')

x = pt.dmatrix("y")
y = pt.dvector("z")

out = np.exp(2 * x * y + y)

x_val = np.random.randn(200, 500)
y_val = np.random.randn(500)

func = pytensor.function([x, y], out, mode="NUMBA")
out = func(x_val, y_val)
np.testing.assert_allclose(np.exp(2 * x_val * y_val + y_val), out)

%timeit func(x_val, y_val)

# Before PR
712 µs ± 6.76 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

# After PR
189 µs ± 2.4 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```