@aseyboldt 2022-12-21
I think I finally got there, tests might just pass this time :-)