@ricardoV94 2022-12-21
I'll try to review soon. In the meantime do you want to add 1 or 2 benchmark tests? https://github.com/pymc-devs/pytensor/pull/139