@twiecki 2022-12-21
@aseyboldt Any more timing results?