@aseyboldt 2022-12-22
@twiecki The above benchmark hasn't really changed.
Hard to say how this impacts real models, for the radon model in the nutpie example this currently reduces my compile time from ~15s to ~12s (lots of noise...) and logp eval time from 12μs to 10μs (much less noise).
I would expect bigger wins for models with more data, and also more wins once we get this merged with the loop fusion and move the sums into the elemwise op.

@ricardoV94 added the code from the description as benchmark.