@ricardoV94 2023-01-03
LGTM, not happy with having to write low-level code that seems as complex if not more than our C implementation 🗡️ 

Left some small questions/comments above