@twiecki 2023-01-03
Yeah, I'd also feel more comfortable if we proved the benefit a bit more of adding this much complexity.