@aseyboldt 2023-01-03
The docs failure looks like it is because of the 6.0 release, which dropped py37 support: https://www.sphinx-doc.org/en/master/changes.html#id9
(now #165)

@twiecki 
> Yeah, I'd also feel more comfortable if we proved the benefit a bit more of adding this much complexity.

Does that post above convince you already?
I can see the maintenance problem, it's just that I don't see alternatives that are much better...