@twiecki 2023-01-04
> Does that post above convince you already?

So it's not just faster but actually allowing us to do things we couldn't do, i.e. reductions during elemwise. In that case I'm convinced.